finish tests for refactored atc code
